refactor: 최신 develop기준 리베이스, 카프카 메시지 생성 시 async 추가
