+++ src/main/java/cc/backend/notice/service/eventListener/ApproveCommitEventListener.java
@@ -1,9 +1,10 @@
-package cc.backend.notice.service;
+package cc.backend.notice.service.eventListener;
 
 import cc.backend.kafka.event.approvalShowEvent.ApprovalShowEvent;
 import cc.backend.kafka.event.approvalShowEvent.ApprovalShowProducer;
 import cc.backend.notice.event.ApproveCommitEvent;
 import lombok.RequiredArgsConstructor;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
@@ -13,6 +14,7 @@ import org.springframework.transaction.event.TransactionalEventListener;
 public class ApproveCommitEventListener {
     private final ApprovalShowProducer approvalShowProducer;
 
+    @Async
     @TransactionalEventListener (phase = TransactionPhase.AFTER_COMMIT)
     public void onApproveCommit(ApproveCommitEvent event) {
